+++ adb/console.c
