+++ adb/file_sync_client.c
@@ -19,14 +19,13 @@
 #include <string.h>
 #include <errno.h>
 #include <sys/stat.h>
-//#include <sys/time.h>
+#include <sys/time.h>
 #include <time.h>
-#include <direct.h>
+#include <dirent.h>
 #include <limits.h>
 #include <sys/types.h>
 #include <zipfile/zipfile.h>
-//#include <utime.h>
-#include <sys/utime.h>
+#include <utime.h>
 
 #include "sysdeps.h"
 #include "adb.h"
@@ -36,31 +35,6 @@
 
 static unsigned long long total_bytes;
 static long long start_time;
-
-
-
-int gettimeofday(struct timeval *tp, void *tzp)
-{
-    time_t clock;
-    struct tm tm;
-    SYSTEMTIME wtm;
-
-    GetLocalTime(&wtm);
-    tm.tm_year     = wtm.wYear - 1900;
-    tm.tm_mon     = wtm.wMonth - 1;
-    tm.tm_mday     = wtm.wDay;
-    tm.tm_hour     = wtm.wHour;
-    tm.tm_min     = wtm.wMinute;
-    tm.tm_sec     = wtm.wSecond;
-    tm. tm_isdst    = -1;
-    clock = mktime(&tm);
-    tp->tv_sec = clock;
-    tp->tv_usec = wtm.wMilliseconds * 1000;
-
-    return (0);
-}
-
-
 
 static long long NOW()
 {
@@ -265,12 +239,8 @@ static int write_data_file(int fd, const char *path, syncsendbuf *sbuf, int show
 
     if (show_progress) {
         // Determine local file size.
-		FILE *stream;
-		stream = fopen(path, "r");
-
         struct stat st;
-		//if (fstat(lfd, &st)) {
-		if (fstat(fileno(stream), &st)) {
+        if (fstat(lfd, &st)) {
             fprintf(stderr,"cannot stat '%s': %s\n", path, strerror(errno));
             return -1;
         }
@@ -339,7 +309,9 @@ static int write_data_buffer(int fd, char* file_buffer, int size, syncsendbuf *s
     return err;
 }
 
-#ifdef HAVE_SYMLINKS
+#if defined(_WIN32)
+extern int write_data_link(int fd, const char *path, syncsendbuf *sbuf) __attribute__((error("no symlinks on Windows")));
+#else
 static int write_data_link(int fd, const char *path, syncsendbuf *sbuf)
 {
     int len, ret;
@@ -365,7 +337,7 @@ static int write_data_link(int fd, const char *path, syncsendbuf *sbuf)
 #endif
 
 static int sync_send(int fd, const char *lpath, const char *rpath,
-                     unsigned mtime, mode_t mode, int verifyApk, int show_progress)
+                     unsigned mtime, mode_t mode, int show_progress)
 {
     syncmsg msg;
     int len, r;
@@ -379,63 +351,6 @@ static int sync_send(int fd, const char *lpath, const char *rpath,
 
     snprintf(tmp, sizeof(tmp), ",%d", mode);
     r = strlen(tmp);
-
-    if (verifyApk) {
-        int lfd;
-        zipfile_t zip;
-        zipentry_t entry;
-        int amt;
-
-        // if we are transferring an APK file, then sanity check to make sure
-        // we have a real zip file that contains an AndroidManifest.xml
-        // this requires that we read the entire file into memory.
-        lfd = adb_open(lpath, O_RDONLY);
-        if(lfd < 0) {
-            fprintf(stderr,"cannot open '%s': %s\n", lpath, strerror(errno));
-            return -1;
-        }
-
-        size = adb_lseek(lfd, 0, SEEK_END);
-        if (size == -1 || -1 == adb_lseek(lfd, 0, SEEK_SET)) {
-            fprintf(stderr, "error seeking in file '%s'\n", lpath);
-            adb_close(lfd);
-            return 1;
-        }
-
-        file_buffer = (char *)malloc(size);
-        if (file_buffer == NULL) {
-            fprintf(stderr, "could not allocate buffer for '%s'\n",
-                    lpath);
-            adb_close(lfd);
-            return 1;
-        }
-        amt = adb_read(lfd, file_buffer, size);
-        if (amt != size) {
-            fprintf(stderr, "error reading from file: '%s'\n", lpath);
-            adb_close(lfd);
-            free(file_buffer);
-            return 1;
-        }
-
-        adb_close(lfd);
-
-        zip = init_zipfile(file_buffer, size);
-        if (zip == NULL) {
-            fprintf(stderr, "file '%s' is not a valid zip file\n",
-                    lpath);
-            free(file_buffer);
-            return 1;
-        }
-
-        entry = lookup_zipentry(zip, "AndroidManifest.xml");
-        release_zipfile(zip);
-        if (entry == NULL) {
-            fprintf(stderr, "file '%s' does not contain AndroidManifest.xml\n",
-                    lpath);
-            free(file_buffer);
-            return 1;
-        }
-    }
 
     msg.req.id = ID_SEND;
     msg.req.namelen = htoll(len + r);
@@ -451,10 +366,8 @@ static int sync_send(int fd, const char *lpath, const char *rpath,
         free(file_buffer);
     } else if (S_ISREG(mode))
         write_data_file(fd, lpath, sbuf, show_progress);
-#ifdef HAVE_SYMLINKS
     else if (S_ISLNK(mode))
         write_data_link(fd, lpath, sbuf);
-#endif
     else
         goto fail;
 
@@ -489,10 +402,10 @@ fail:
     return -1;
 }
 
-static int mkdirs(char *name)
+static int mkdirs(const char *name)
 {
     int ret;
-    char *x = name + 1;
+    char *x = (char *)name + 1;
 
     for(;;) {
         x = adb_dirstart(x);
@@ -554,7 +467,7 @@ int sync_recv(int fd, const char *rpath, const char *lpath, int show_progress)
 
     if((id == ID_DATA) || (id == ID_DONE)) {
         adb_unlink(lpath);
-		mkdirs((char *)lpath);
+        mkdirs(lpath);
         lfd = adb_creat(lpath, 0644);
         if(lfd < 0) {
             fprintf(stderr,"cannot create '%s': %s\n", lpath, strerror(errno));
@@ -672,7 +585,7 @@ copyinfo *mkcopyinfo(const char *spath, const char *dpath,
     int ssize = slen + nlen + 2;
     int dsize = dlen + nlen + 2;
 
-	copyinfo *ci = (copyinfo *)malloc(sizeof(copyinfo) + ssize + dsize);
+    copyinfo *ci = malloc(sizeof(copyinfo) + ssize + dsize);
     if(ci == 0) {
         fprintf(stderr,"out of memory\n");
         abort();
@@ -696,33 +609,23 @@ copyinfo *mkcopyinfo(const char *spath, const char *dpath,
 static int local_build_list(copyinfo **filelist,
                             const char *lpath, const char *rpath)
 {
- //   DIR *d;
- //   struct dirent *de;
-	struct _finddata_t c_file;
-	intptr_t   hFile; 
+    DIR *d;
+    struct dirent *de;
     struct stat st;
     copyinfo *dirlist = 0;
     copyinfo *ci, *next;
 
 //    fprintf(stderr,"local_build_list('%s','%s')\n", lpath, rpath);
 
-   // d = opendir(lpath);
-//	if(d == 0)
-	if(_chdir(lpath)) 
-     {
+    d = opendir(lpath);
+    if(d == 0) {
         fprintf(stderr,"cannot open '%s': %s\n", lpath, strerror(errno));
         return -1;
     }
-	else
-	{
-		hFile = _findfirst("*.*", &c_file); 
-	}
 
-   // while((de = readdir(d))) 
-	do 
-	{
-        char stat_path[MAX_PATH];
-        char *name = c_file.name;
+    while((de = readdir(d))) {
+        char stat_path[PATH_MAX];
+        char *name = de->d_name;
 
         if(name[0] == '.') {
             if(name[1] == 0) continue;
@@ -733,42 +636,41 @@ static int local_build_list(copyinfo **filelist,
          * We could use d_type if HAVE_DIRENT_D_TYPE is defined, but reiserfs
          * always returns DT_UNKNOWN, so we just use stat() for all cases.
          */
-        if (strlen(lpath) + strlen(c_file.name) + 1 > sizeof(stat_path))
-		{
-			_findnext(hFile, &c_file);
+        if (strlen(lpath) + strlen(de->d_name) + 1 > sizeof(stat_path))
             continue;
-		}
         strcpy(stat_path, lpath);
-        strcat(stat_path, c_file.name);
-        stat(stat_path, &st);
+        strcat(stat_path, de->d_name);
 
-        if (S_IFDIR&st.st_mode) {
-            ci = mkcopyinfo(lpath, rpath, name, 1);
-            ci->next = dirlist;
-            dirlist = ci;
-        } else {
-            ci = mkcopyinfo(lpath, rpath, name, 0);
-            if(lstat(ci->src, &st)) {
-             //   closedir(d);
-				_findclose(hFile);
-                fprintf(stderr,"cannot stat '%s': %s\n", ci->src, strerror(errno));
-                return -1;
-            }
-            if(!S_ISREG(st.st_mode) && !S_ISLNK(st.st_mode)) {
-                fprintf(stderr, "skipping special file '%s'\n", ci->src);
-                free(ci);
+        if(!lstat(stat_path, &st)) {
+            if (S_ISDIR(st.st_mode)) {
+                ci = mkcopyinfo(lpath, rpath, name, 1);
+                ci->next = dirlist;
+                dirlist = ci;
             } else {
-                ci->time = st.st_mtime;
-                ci->mode = st.st_mode;
-                ci->size = st.st_size;
-                ci->next = *filelist;
-                *filelist = ci;
+                ci = mkcopyinfo(lpath, rpath, name, 0);
+                if(lstat(ci->src, &st)) {
+                    fprintf(stderr,"cannot stat '%s': %s\n", ci->src, strerror(errno));
+                    free(ci);
+                    closedir(d);
+                    return -1;
+                }
+                if(!S_ISREG(st.st_mode) && !S_ISLNK(st.st_mode)) {
+                    fprintf(stderr, "skipping special file '%s'\n", ci->src);
+                    free(ci);
+                } else {
+                    ci->time = st.st_mtime;
+                    ci->mode = st.st_mode;
+                    ci->size = st.st_size;
+                    ci->next = *filelist;
+                    *filelist = ci;
+                }
             }
+        } else {
+            fprintf(stderr, "cannot lstat '%s': %s\n",stat_path , strerror(errno));
         }
-    }while(_findnext(hFile, &c_file) == 0);
+    }
 
-  //  closedir(d);
-	_findclose(hFile); 
+    closedir(d);
 
     for(ci = dirlist; ci != 0; ci = next) {
         next = ci->next;
@@ -790,14 +692,14 @@ static int copy_local_dir_remote(int fd, const char *lpath, const char *rpath, i
     if((lpath[0] == 0) || (rpath[0] == 0)) return -1;
     if(lpath[strlen(lpath) - 1] != '/') {
         int  tmplen = strlen(lpath)+2;
-        char *tmp =(char *) malloc(tmplen);
+        char *tmp = malloc(tmplen);
         if(tmp == 0) return -1;
         snprintf(tmp, tmplen, "%s/",lpath);
         lpath = tmp;
     }
     if(rpath[strlen(rpath) - 1] != '/') {
         int tmplen = strlen(rpath)+2;
-        char *tmp =(char *) malloc(tmplen);
+        char *tmp = malloc(tmplen);
         if(tmp == 0) return -1;
         snprintf(tmp, tmplen, "%s/",rpath);
         rpath = tmp;
@@ -830,7 +732,7 @@ static int copy_local_dir_remote(int fd, const char *lpath, const char *rpath, i
         if(ci->flag == 0) {
             fprintf(stderr,"%spush: %s -> %s\n", listonly ? "would " : "", ci->src, ci->dst);
             if(!listonly &&
-               sync_send(fd, ci->src, ci->dst, ci->time, ci->mode, 0 /* no verify APK */,
+               sync_send(fd, ci->src, ci->dst, ci->time, ci->mode,
                          0 /* no show progress */)) {
                 return 1;
             }
@@ -849,7 +751,7 @@ static int copy_local_dir_remote(int fd, const char *lpath, const char *rpath, i
 }
 
 
-int do_sync_push(const char *lpath, const char *rpath, int verifyApk, int show_progress)
+int do_sync_push(const char *lpath, const char *rpath, int show_progress)
 {
     struct stat st;
     unsigned mode;
@@ -890,15 +792,13 @@ int do_sync_push(const char *lpath, const char *rpath, int verifyApk, int show_p
                 name++;
             }
             int  tmplen = strlen(name) + strlen(rpath) + 2;
-			char *tmp = (char *)malloc(strlen(name) + strlen(rpath) + 2);
+            char *tmp = malloc(strlen(name) + strlen(rpath) + 2);
             if(tmp == 0) return 1;
             snprintf(tmp, tmplen, "%s/%s", rpath, name);
             rpath = tmp;
         }
-		std::string utfrpath;
-		GBK_to_UTF8(rpath, strlen(rpath), utfrpath);
-		BEGIN();
-		if (sync_send(fd, lpath, utfrpath.c_str(), st.st_mtime, st.st_mode, verifyApk, show_progress)) {
+        BEGIN();
+        if(sync_send(fd, lpath, rpath, st.st_mtime, st.st_mode, show_progress)) {
             return 1;
         } else {
             END();
@@ -982,7 +882,7 @@ static int remote_build_list(int syncfd, copyinfo **filelist,
 
 static int set_time_and_mode(const char *lpath, unsigned int time, unsigned int mode)
 {
-    utimbuf times = { time, time };
+    struct utimbuf times = { time, time };
     int r1 = utime(lpath, &times);
 
     /* use umask for permissions */
@@ -990,7 +890,22 @@ static int set_time_and_mode(const char *lpath, unsigned int time, unsigned int 
     umask(mask);
     int r2 = chmod(lpath, mode & ~mask);
 
-	return r1 ? 0 : r2;
+    return r1 ? : r2;
+}
+
+/* Return a copy of the path string with / appended if needed */
+static char *add_slash_to_path(const char *path)
+{
+    if (path[strlen(path) - 1] != '/') {
+        size_t len = strlen(path) + 2;
+        char *path_with_slash = malloc(len);
+        if (path_with_slash == NULL)
+            return NULL;
+        snprintf(path_with_slash, len, "%s/", path);
+        return path_with_slash;
+    } else {
+        return strdup(path);
+    }
 }
 
 static int copy_remote_dir_local(int fd, const char *rpath, const char *lpath,
@@ -1000,28 +915,32 @@ static int copy_remote_dir_local(int fd, const char *rpath, const char *lpath,
     copyinfo *ci, *next;
     int pulled = 0;
     int skipped = 0;
+    char *rpath_clean = NULL;
+    char *lpath_clean = NULL;
+    int ret = 0;
+
+    if (rpath[0] == '\0' || lpath[0] == '\0') {
+        ret = -1;
+        goto finish;
+    }
 
     /* Make sure that both directory paths end in a slash. */
-    if (rpath[0] == 0 || lpath[0] == 0) return -1;
-    if (rpath[strlen(rpath) - 1] != '/') {
-        int  tmplen = strlen(rpath) + 2;
-        char *tmp =(char *) malloc(tmplen);
-        if (tmp == 0) return -1;
-        snprintf(tmp, tmplen, "%s/", rpath);
-        rpath = tmp;
+    rpath_clean = add_slash_to_path(rpath);
+    if (!rpath_clean) {
+        ret = -1;
+        goto finish;
     }
-    if (lpath[strlen(lpath) - 1] != '/') {
-        int  tmplen = strlen(lpath) + 2;
-        char *tmp = (char *)malloc(tmplen);
-        if (tmp == 0) return -1;
-        snprintf(tmp, tmplen, "%s/", lpath);
-        lpath = tmp;
+    lpath_clean = add_slash_to_path(lpath);
+    if (!lpath_clean) {
+        ret = -1;
+        goto finish;
     }
 
-    fprintf(stderr, "pull: building file list...\n");
     /* Recursively build the list of files to copy. */
-    if (remote_build_list(fd, &filelist, rpath, lpath)) {
-        return -1;
+    fprintf(stderr, "pull: building file list...\n");
+    if (remote_build_list(fd, &filelist, rpath_clean, lpath_clean)) {
+        ret = -1;
+        goto finish;
     }
 
     for (ci = filelist; ci != 0; ci = next) {
@@ -1029,11 +948,13 @@ static int copy_remote_dir_local(int fd, const char *rpath, const char *lpath,
         if (ci->flag == 0) {
             fprintf(stderr, "pull: %s -> %s\n", ci->src, ci->dst);
             if (sync_recv(fd, ci->src, ci->dst, 0 /* no show progress */)) {
-                return 1;
+                ret = -1;
+                goto finish;
             }
 
             if (copy_attrs && set_time_and_mode(ci->dst, ci->time, ci->mode)) {
-               return 1;
+                ret = -1;
+                goto finish;
             }
             pulled++;
         } else {
@@ -1046,14 +967,14 @@ static int copy_remote_dir_local(int fd, const char *rpath, const char *lpath,
             pulled, (pulled == 1) ? "" : "s",
             skipped, (skipped == 1) ? "" : "s");
 
-    return 0;
+finish:
+    free(lpath_clean);
+    free(rpath_clean);
+    return ret;
 }
 
 int do_sync_pull(const char *rpath, const char *lpath, int show_progress, int copy_attrs)
 {
-	std::string utfrpath;
-	GBK_to_UTF8(rpath, strlen(rpath), utfrpath);
-
     unsigned mode, time;
     struct stat st;
 
@@ -1065,7 +986,7 @@ int do_sync_pull(const char *rpath, const char *lpath, int show_progress, int co
         return 1;
     }
 
-    if(sync_readtime(fd, utfrpath.c_str(), &time, &mode)) {
+    if(sync_readtime(fd, rpath, &time, &mode)) {
         return 1;
     }
     if(mode == 0) {
@@ -1086,19 +1007,18 @@ int do_sync_pull(const char *rpath, const char *lpath, int show_progress, int co
                     name++;
                 }
                 int  tmplen = strlen(name) + strlen(lpath) + 2;
-                char *tmp = (char *)malloc(tmplen);
+                char *tmp = malloc(tmplen);
                 if(tmp == 0) return 1;
                 snprintf(tmp, tmplen, "%s/%s", lpath, name);
                 lpath = tmp;
             }
         }
-
         BEGIN();
-		if (sync_recv(fd, utfrpath.c_str(), lpath, show_progress)) {
+        if (sync_recv(fd, rpath, lpath, show_progress)) {
             return 1;
         } else {
-			if (copy_attrs && set_time_and_mode(lpath, time, mode))
-				return 1;
+            if (copy_attrs && set_time_and_mode(lpath, time, mode))
+                return 1;
             END();
             sync_quit(fd);
             return 0;
