+++ adb/transport_usb.c
@@ -18,37 +18,10 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "sysdeps.h"
+#include <sysdeps.h>
 
 #define  TRACE_TAG  TRACE_TRANSPORT
 #include "adb.h"
-
-#if ADB_HOST
-#include "usb_vendors.h"
-#endif
-
-#ifdef HAVE_BIG_ENDIAN
-#define H4(x)	(((x) & 0xFF000000) >> 24) | (((x) & 0x00FF0000) >> 8) | (((x) & 0x0000FF00) << 8) | (((x) & 0x000000FF) << 24)
-static inline void fix_endians(apacket *p)
-{
-    p->msg.command     = H4(p->msg.command);
-    p->msg.arg0        = H4(p->msg.arg0);
-    p->msg.arg1        = H4(p->msg.arg1);
-    p->msg.data_length = H4(p->msg.data_length);
-    p->msg.data_check  = H4(p->msg.data_check);
-    p->msg.magic       = H4(p->msg.magic);
-}
-unsigned host_to_le32(unsigned n)
-{
-    return H4(n);
-}
-#else
-#define fix_endians(p) do {} while (0)
-unsigned host_to_le32(unsigned n)
-{
-    return n;
-}
-#endif
 
 static int remote_read(apacket *p, atransport *t)
 {
@@ -56,8 +29,6 @@ static int remote_read(apacket *p, atransport *t)
         D("remote usb: read terminated (message)\n");
         return -1;
     }
-
-    fix_endians(p);
 
     if(check_header(p)) {
         D("remote usb: check_header failed\n");
@@ -82,8 +53,6 @@ static int remote_read(apacket *p, atransport *t)
 static int remote_write(apacket *p, atransport *t)
 {
     unsigned size = p->msg.data_length;
-
-    fix_endians(p);
 
     if(usb_write(t->usb, &p->msg, sizeof(amessage))) {
         D("remote usb: 1 - write terminated\n");
@@ -131,18 +100,6 @@ void init_usb_transport(atransport *t, usb_handle *h, int state)
 #if ADB_HOST
 int is_adb_interface(int vid, int pid, int usb_class, int usb_subclass, int usb_protocol)
 {
-    unsigned i;
-    for (i = 0; i < vendorIdCount; i++) {
-        if (vid == vendorIds[i]) {
-            if (usb_class == ADB_CLASS && usb_subclass == ADB_SUBCLASS &&
-                    usb_protocol == ADB_PROTOCOL) {
-                return 1;
-            }
-
-            return 0;
-        }
-    }
-
-    return 0;
+    return (usb_class == ADB_CLASS && usb_subclass == ADB_SUBCLASS && usb_protocol == ADB_PROTOCOL);
 }
 #endif
